+++ test/snippet/argument_parser/validators_input_file.cpp
@@ -10,7 +10,7 @@ int main(int argc, const char ** argv)
     std::filesystem::path myfile{};
 
     myparser.add_option(myfile,'f',"file","The input file containing the sequences.",
-                        seqan3::option_spec::DEFAULT, seqan3::input_file_validator{{"fa","fasta"}});
+                        seqan3::option_spec::defaulted, seqan3::input_file_validator{{"fa","fasta"}});
     //! [validator_call]
 
     // an exception will be thrown if the user specifies a filename